While testing the serial communication I created two new files serialTest.c and serialTest.h
You can ignore if you want but it is convenient for testing. Also a new file def.h is now provided with common definitions (#define statements) so that we do not have to include rev3.h in all our files. The define statements in def.h are for better readability of the code, we should merge towards using these statements, but not actively change the code.

FFG